requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
psycopg2-binary
pandas
google-genai
python-dotenv
openpyxl
pypdf
tabulate
werkzeug
fpdf
